requirements.txt
flask
google-generativeai
python-dotenv
openai
waitress
